trading_backend/app/services/web_scraping/account_discovery.py
"""Account discovery service for social media platforms."""
import logging
from typing import List, Dict, Set, Optional
import asyncio
from datetime import datetime, timedelta
import json

logger = logging.getLogger(__name__)

class AccountDiscoveryService:
    """Service for discovering and managing influential accounts."""

    def __init__(self):
        # Initial seed accounts (influential figures and institutions)
        self.seed_accounts = {
            'twitter': [
                'elonmusk',        # Elon Musk
                'cz_binance',      # Binance CEO
                'VitalikButerin',  # Ethereum founder
                'saylor',          # Michael Saylor
                'CryptoCapo_',     # Crypto analyst
                'DocumentingBTC',  # Bitcoin news
                'BitcoinMagazine', # Bitcoin Magazine
                'Cointelegraph',   # Crypto news
                'CoinDesk',        # Crypto news
                'TheBlock__'       # Crypto research
            ],
            'youtube': [
                'BitcoinMagazine',
                'CoinBureau',
                'Cointelegraph',
                'CryptoDaily',
                'DataDash'
            ]
        }

        # Store discovered accounts with their metrics
        self.discovered_accounts: Dict[str, Dict[str, Dict]] = {
            'twitter': {},
            'youtube': {}
        }

        # Track last update time for rate limiting
        self.last_update: Dict[str, datetime] = {
            'twitter': datetime.min,
            'youtube': datetime.min
        }

        # Rate limiting settings (requests per hour)
        self.rate_limits = {
            'twitter': 300,  # Twitter API rate limit
            'youtube': 100   # YouTube API rate limit
        }

    async def initialize(self) -> None:
        """Initialize the service and load any saved account data."""
        try:
            # Load previously discovered accounts if available
            for platform in self.discovered_accounts:
                try:
                    with open(f'data/{platform}_accounts.json', 'r') as f:
                        self.discovered_accounts[platform] = json.load(f)
                except FileNotFoundError:
                    logger.info(f"No existing account data for {platform}")

            # Initialize with seed accounts
            await self.update_all_platforms()
        except Exception as e:
            logger.error(f"Error initializing account discovery: {str(e)}")
            raise

    async def update_all_platforms(self) -> None:
        """Update account discoveries for all platforms."""
        tasks = []
        for platform in self.seed_accounts:
            if self._can_update(platform):
                tasks.append(self.discover_accounts(platform))

        if tasks:
            await asyncio.gather(*tasks)

    def _can_update(self, platform: str) -> bool:
        """Check if we can update based on rate limits."""
        now = datetime.now()
        time_diff = now - self.last_update[platform]

        # Allow update if more than 1 hour has passed
        if time_diff > timedelta(hours=1):
            self.last_update[platform] = now
            return True
        return False

    async def discover_accounts(self, platform: str) -> None:
        """Discover new accounts based on interactions with seed accounts."""
        try:
            if platform == 'twitter':
                await self._discover_twitter_accounts()
            elif platform == 'youtube':
                await self._discover_youtube_accounts()
        except Exception as e:
            logger.error(f"Error discovering accounts for {platform}: {str(e)}")

    async def _discover_twitter_accounts(self) -> None:
        """Discover influential Twitter accounts."""
        # Implementation will be added when Twitter scraper is ready
        pass

    async def _discover_youtube_accounts(self) -> None:
        """Discover influential YouTube accounts."""
        # Implementation will be added when YouTube scraper is ready
        pass

    def get_top_accounts(self, platform: str, limit: int = 10) -> List[str]:
        """Get top accounts for a platform based on influence metrics."""
        try:
            accounts = self.discovered_accounts.get(platform, {})
            sorted_accounts = sorted(
                accounts.items(),
                key=lambda x: (
                    x[1].get('engagement_rate', 0),
                    x[1].get('followers', 0)
                ),
                reverse=True
            )
            return [account[0] for account in sorted_accounts[:limit]]
        except Exception as e:
            logger.error(f"Error getting top accounts: {str(e)}")
            return []

    def add_seed_account(self, platform: str, username: str) -> bool:
        """Add a new seed account for discovery."""
        try:
            if platform in self.seed_accounts:
                if username not in self.seed_accounts[platform]:
                    self.seed_accounts[platform].append(username)
                    return True
            return False
        except Exception as e:
            logger.error(f"Error adding seed account: {str(e)}")
            return False


    async def save_discovered_accounts(self) -> None:
        """Save discovered accounts to persistent storage."""
        try:
            for platform in self.discovered_accounts:
                with open(f'data/{platform}_accounts.json', 'w') as f:
                    json.dump(self.discovered_accounts[platform], f)
        except Exception as e:
            logger.error(f"Error saving discovered accounts: {str(e)}")

    def get_account_metrics(self, platform: str, username: str) -> Optional[Dict]:
        """Get metrics for a specific account."""
        try:
            return self.discovered_accounts.get(platform, {}).get(username)
        except Exception as e:
            logger.error(f"Error getting account metrics: {str(e)}")
            return None
